fix: Move policy and simulation frames to separate folders
