+++ app/src/flows/registry.ts
@@ -1,18 +1,18 @@
 import { PolicyViewFlow } from '@/flows/policyViewFlow';
-import PolicyCreationFrame from '@/frames/PolicyCreationFrame';
-import PolicyParameterSelectorFrame from '@/frames/PolicyParameterSelectorFrame';
-import PolicySubmitFrame from '@/frames/PolicySubmitFrame';
+import PolicyCreationFrame from '@/frames/policy/PolicyCreationFrame';
+import PolicyParameterSelectorFrame from '@/frames/policy/PolicyParameterSelectorFrame';
+import PolicySubmitFrame from '@/frames/policy/PolicySubmitFrame';
 import GeographicConfirmationFrame from '@/frames/population/GeographicConfirmationFrame';
 import HouseholdBuilderFrame from '@/frames/population/HouseholdBuilderFrame';
 import SelectGeographicScopeFrame from '@/frames/population/SelectGeographicScopeFrame';
 import SetPopulationLabelFrame from '@/frames/population/SetPopulationLabelFrame';
-import SimulationCreationFrame from '@/frames/SimulationCreationFrame';
-import SimulationSelectExistingPolicyFrame from '@/frames/SimulationSelectExistingPolicyFrame';
-import SimulationSelectExistingPopulationFrame from '@/frames/SimulationSelectExistingPopulationFrame';
-import SimulationSetupFrame from '@/frames/SimulationSetupFrame';
-import SimulationSetupPolicyFrame from '@/frames/SimulationSetupPolicyFrame';
-import SimulationSetupPopulationFrame from '@/frames/SimulationSetupPopulationFrame';
-import SimulationSubmitFrame from '@/frames/SimulationSubmitFrame';
+import SimulationCreationFrame from '@/frames/simulation/SimulationCreationFrame';
+import SimulationSelectExistingPolicyFrame from '@/frames/simulation/SimulationSelectExistingPolicyFrame';
+import SimulationSelectExistingPopulationFrame from '@/frames/simulation/SimulationSelectExistingPopulationFrame';
+import SimulationSetupFrame from '@/frames/simulation/SimulationSetupFrame';
+import SimulationSetupPolicyFrame from '@/frames/simulation/SimulationSetupPolicyFrame';
+import SimulationSetupPopulationFrame from '@/frames/simulation/SimulationSetupPopulationFrame';
+import SimulationSubmitFrame from '@/frames/simulation/SimulationSubmitFrame';
 import PoliciesPage from '@/pages/Policies.page';
 import PopulationsPage from '@/pages/Populations.page';
 import SimulationsPage from '@/pages/Simulations.page';
